Refactor tests to avoid repeating the same use cases
