+++ pipeline/merger/src/test/scala/uk/ac/wellcome/platform/merger/rules/MergerRuleTest.scala
@@ -11,6 +11,41 @@ import uk.ac.wellcome.models.work.internal.{
 import uk.ac.wellcome.platform.merger.model.MergedWork
 
 class MergerRuleTest extends FunSpec with WorksGenerators with Matchers {
+  it("merges a pair of works in a list") {
+    val mergerRule = new MergerRule with Partitioner with WorkPairMerger {
+      override protected def partitionWorks(
+                                             works: Seq[BaseWork]): Option[Partition] =
+        Some(
+          Partition(
+            PotentialMergedWork(
+              works.head.asInstanceOf[UnidentifiedWork],
+              works.tail.head.asInstanceOf[UnidentifiedWork]
+            ),
+            works.tail.tail
+          )
+        )
+      override protected def mergeAndRedirectWorkPair(
+                                                       firstWork: UnidentifiedWork,
+                                                       secondWork: TransformedBaseWork): Option[MergedWork] =
+        Some(
+          MergedWork(
+            firstWork,
+            createUnidentifiedRedirectedWork(secondWork, firstWork)
+          )
+        )
+    }
+    val works = createUnidentifiedWorks(5)
+
+    val expectedMergedWork =
+      works.head.withData(data => data.copy(merged = true))
+    val expectedRedirectedWork = UnidentifiedRedirectedWork(
+      version = 1,
+      sourceIdentifier = works.tail.head.sourceIdentifier,
+      redirect = IdentifiableRedirect(works.head.sourceIdentifier))
+    val expectedWorks = expectedMergedWork +: expectedRedirectedWork +: works.tail.tail
+    mergerRule.mergeAndRedirectWorks(works) shouldBe expectedWorks
+  }
+
   it("returns the works unchanged if the list cannot be partitioned") {
     val mergerRule = new MergerRule with Partitioner with WorkPairMerger {
       override protected def partitionWorks(
@@ -45,42 +80,6 @@ class MergerRuleTest extends FunSpec with WorksGenerators with Matchers {
 
     val works = createUnidentifiedWorks(5)
     mergerRule.mergeAndRedirectWorks(works) shouldBe works
-  }
-
-  it("sets the merged flag of merged works") {
-    val mergerRule = new MergerRule with Partitioner with WorkPairMerger {
-      override protected def partitionWorks(
-        works: Seq[BaseWork]): Option[Partition] =
-        Some(
-          Partition(
-            PotentialMergedWork(
-              works.head.asInstanceOf[UnidentifiedWork],
-              works.tail.head.asInstanceOf[UnidentifiedWork]
-            ),
-            works.tail.tail
-          )
-        )
-      override protected def mergeAndRedirectWorkPair(
-        firstWork: UnidentifiedWork,
-        secondWork: TransformedBaseWork): Option[MergedWork] =
-        Some(
-          MergedWork(
-            firstWork,
-            createUnidentifiedRedirectedWork(secondWork, firstWork)
-          )
-        )
-    }
-
-    val works = createUnidentifiedWorks(5)
-
-    val expectedMergedWork =
-      works.head.withData(data => data.copy(merged = true))
-    val expectedRedirectedWork = UnidentifiedRedirectedWork(
-      version = 1,
-      sourceIdentifier = works.tail.head.sourceIdentifier,
-      redirect = IdentifiableRedirect(works.head.sourceIdentifier))
-    val expectedWorks = expectedMergedWork +: expectedRedirectedWork +: works.tail.tail
-    mergerRule.mergeAndRedirectWorks(works) shouldBe expectedWorks
   }
 
 }
